refactor(yaml): split loader.ts into state, scanner and scalar modules
loader.ts had grown to ~1900 lines. Extract the parts that have no cycle
back into the composer:

- state.ts   — the State cursor, Snapshot, speculation undo, throw/warn
- scanner.ts — character predicates and low-level cursor moves
- scalars.ts — the four scalar readers (leaf parsers, verified acyclic)

loader.ts keeps composeNode and the collection readers, which are
mutually recursive and would only gain an import cycle by being split.
It drops from 1867 to 1162 lines; dependencies run strictly one way
(state <- scanner <- scalars <- loader).

No behaviour change: 402/402 conformance and all 117 tests still pass,
and packem bundles the modules into a single dist/index.js so there is no
runtime cost.
