+++ packages/data-manipulation/yaml/src/parser/loader.ts
@@ -5,7 +5,7 @@
  * reassignment), `charCodeAt` is used deliberately for byte-level scanning with
  * a `0` EOF sentinel, and YAML's `null` is a first-class value in the output.
  */
-/* eslint-disable no-bitwise */
+
 /* eslint-disable no-plusplus */
 /* eslint-disable no-cond-assign */
 /* eslint-disable no-useless-assignment */
@@ -30,355 +30,24 @@
  * no intermediate concrete syntax tree on the default path).
  */
 
-import { YAMLParseError, YAMLWarning } from "../errors";
+import { YAMLParseError } from "../errors";
 import { resolveExplicitTag, resolveScalarValue } from "../schema/resolve-scalar";
 import type { ParseOptions } from "../types";
+import { readBlockScalar, readDoubleQuotedScalar, readPlainScalar, readSingleQuotedScalar } from "./scalars";
+import { isEol, isFlowIndicator, isWhiteSpace, isWsOrEol, readLineBreak, skipSeparationSpace, testDocumentSeparator } from "./scanner";
+import type { Snapshot } from "./state";
+import { beginSpeculation, emitWarning, restoreState, rollbackSpeculation, snapshotState, State, throwError } from "./state";
+
+const isPlainObject = (value: unknown): boolean => typeof value === "object" && value !== null && Object.prototype.toString.call(value) === "[object Object]";
 
 const CONTEXT_FLOW_IN = 1;
 const CONTEXT_FLOW_OUT = 2;
 const CONTEXT_BLOCK_IN = 3;
 const CONTEXT_BLOCK_OUT = 4;
 
-const CHOMPING_CLIP = 1;
-const CHOMPING_STRIP = 2;
-const CHOMPING_KEEP = 3;
-
 const MERGE_TAG = "tag:yaml.org,2002:merge";
 
 const YAML_VERSION_RE = /^\d+\.\d+$/;
-
-const isEol = (c: number): boolean => c === 0x0a || c === 0x0d;
-const isWhiteSpace = (c: number): boolean => c === 0x09 || c === 0x20;
-const isWsOrEol = (c: number): boolean => c === 0x09 || c === 0x20 || c === 0x0a || c === 0x0d || c === 0;
-const isFlowIndicator = (c: number): boolean => c === 0x2c || c === 0x5b || c === 0x5d || c === 0x7b || c === 0x7d;
-
-// `# & * ! | > ' " % @ \`` — none of these may begin a plain scalar. A 128-entry
-// lookup table is faster than `Set.has` on the scalar hot path.
-const PLAIN_SCALAR_LEAD_BLOCKERS = new Uint8Array(128);
-
-for (const code of [0x21, 0x22, 0x23, 0x25, 0x26, 0x27, 0x2a, 0x3e, 0x40, 0x60, 0x7c]) {
-    PLAIN_SCALAR_LEAD_BLOCKERS[code] = 1;
-}
-
-const isPlainScalarLeadBlocker = (c: number): boolean => c < 128 && PLAIN_SCALAR_LEAD_BLOCKERS[c] === 1;
-
-const fromHexCode = (c: number): number => {
-    if (c >= 0x30 && c <= 0x39) {
-        return c - 0x30;
-    }
-
-    const lc = c | 0x20;
-
-    if (lc >= 0x61 && lc <= 0x66) {
-        return lc - 0x61 + 10;
-    }
-
-    return -1;
-};
-
-const fromDecimalCode = (c: number): number => {
-    if (c >= 0x30 && c <= 0x39) {
-        return c - 0x30;
-    }
-
-    return -1;
-};
-
-const SIMPLE_ESCAPE: (string | undefined)[] = Array.from({ length: 256 });
-
-SIMPLE_ESCAPE[0x30] = "\0";
-SIMPLE_ESCAPE[0x61] = "\u0007";
-SIMPLE_ESCAPE[0x62] = "\b";
-SIMPLE_ESCAPE[0x74] = "\t";
-SIMPLE_ESCAPE[0x09] = "\t";
-SIMPLE_ESCAPE[0x6e] = "\n";
-SIMPLE_ESCAPE[0x76] = "\v";
-SIMPLE_ESCAPE[0x66] = "\f";
-SIMPLE_ESCAPE[0x72] = "\r";
-SIMPLE_ESCAPE[0x65] = "\u001B";
-SIMPLE_ESCAPE[0x20] = " ";
-SIMPLE_ESCAPE[0x22] = "\u0022";
-SIMPLE_ESCAPE[0x2f] = "/";
-SIMPLE_ESCAPE[0x5c] = "\\";
-SIMPLE_ESCAPE[0x4e] = "\u0085";
-SIMPLE_ESCAPE[0x5f] = "\u00A0";
-SIMPLE_ESCAPE[0x4c] = "\u2028";
-SIMPLE_ESCAPE[0x50] = "\u2029";
-
-const escapedHexLength = (c: number): number => {
-    switch (c) {
-        case 0x55: {
-            return 8;
-        }
-        case 0x75: {
-            return 4;
-        }
-        case 0x78: {
-            return 2;
-        }
-        default: {
-            return 0;
-        }
-    }
-};
-
-/** The three node shapes the composer distinguishes. */
-type NodeKind = "mapping" | "scalar" | "sequence";
-
-const isPlainObject = (value: unknown): boolean => typeof value === "object" && value !== null && Object.prototype.toString.call(value) === "[object Object]";
-
-/** The mutable parser cursor + accumulators. */
-class State {
-    public input: string;
-
-    public length: number;
-
-    public position = 0;
-
-    public line = 0;
-
-    public lineStart = 0;
-
-    public lineIndent = 0;
-
-    // Position of the first tab in the current line's leading white space, or
-    // -1. Tabs are illegal as block indentation; block collections consult this
-    // to reject them. Reset on every line break.
-    public firstTabInLine = -1;
-
-    // Line of a `---` marker that carried content on the same line, or -1. In
-    // strict mode a block collection whose first entry sits on this line is
-    // rejected (`--- a: b`). Reset once the document's top node is composed.
-    public documentMarkerLine = -1;
-
-    public documents: unknown[] = [];
-
-    public tag: string | null = null;
-
-    public anchor: string | null = null;
-
-    public kind: NodeKind | null = null;
-
-    public result: unknown = null;
-
-    public anchorMap = new Map<string, unknown>();
-
-    public tagMap = new Map<string, string>();
-
-    public aliasCount = 0;
-
-    /** Current `composeNode` recursion depth; bounded by `options.maxDepth`. */
-    public depth = 0;
-
-    public readonly options: ParseOptions & Required<Pick<ParseOptions, "duplicateKeys" | "maxAliasCount" | "maxDepth" | "preventProtoPollution" | "strict">>;
-
-    public constructor(input: string, options: ParseOptions) {
-        this.input = input;
-        this.length = input.length;
-        // Spread first, then apply the defaults with `??`. The other order lets a
-        // key that is present but `undefined` — the shape you get from
-        // `parse(src, { maxAliasCount: config.maxAliasCount })` — overwrite the
-        // default, silently disabling the alias limit, the duplicate-key error,
-        // strict mode and the prototype-pollution guard.
-        this.options = {
-            ...options,
-            duplicateKeys: options.duplicateKeys ?? "error",
-            maxAliasCount: options.maxAliasCount ?? 100,
-            maxDepth: options.maxDepth ?? 1000,
-            preventProtoPollution: options.preventProtoPollution ?? true,
-            strict: options.strict ?? true,
-        };
-    }
-}
-
-interface Snapshot {
-    anchor: string | null;
-    firstTabInLine: number;
-    kind: NodeKind | null;
-    line: number;
-    lineIndent: number;
-    lineStart: number;
-    position: number;
-    result: unknown;
-    tag: string | null;
-}
-
-const snapshotState = (state: State): Snapshot => {
-    return {
-        anchor: state.anchor,
-        firstTabInLine: state.firstTabInLine,
-        kind: state.kind,
-        line: state.line,
-        lineIndent: state.lineIndent,
-        lineStart: state.lineStart,
-        position: state.position,
-        result: state.result,
-        tag: state.tag,
-    };
-};
-
-const restoreState = (state: State, snapshot: Snapshot): void => {
-    state.position = snapshot.position;
-    state.line = snapshot.line;
-    state.lineStart = snapshot.lineStart;
-    state.lineIndent = snapshot.lineIndent;
-    state.firstTabInLine = snapshot.firstTabInLine;
-    state.tag = snapshot.tag;
-    state.anchor = snapshot.anchor;
-    state.kind = snapshot.kind;
-    state.result = snapshot.result;
-};
-
-/**
- * State that a speculative parse must be able to undo, beyond the cursor that
- * `Snapshot` already covers: anchors registered by the attempt and the alias
- * budget it consumed.
- */
-interface SpeculationUndo {
-    aliasCount: number;
-    anchorMap: Map<string, unknown>;
-    cursor: Snapshot;
-}
-
-const beginSpeculation = (state: State): SpeculationUndo => {
-    return { aliasCount: state.aliasCount, anchorMap: new Map(state.anchorMap), cursor: snapshotState(state) };
-};
-
-const rollbackSpeculation = (state: State, undo: SpeculationUndo): void => {
-    state.anchorMap = undo.anchorMap;
-    state.aliasCount = undo.aliasCount;
-    restoreState(state, undo.cursor);
-};
-
-const throwError = (state: State, message: string): never => {
-    throw new YAMLParseError(
-        message,
-        {
-            column: state.position - state.lineStart,
-            line: state.line,
-            position: state.position,
-        },
-        state.input,
-    );
-};
-
-const emitWarning = (state: State, message: string): void => {
-    if (state.options.onWarning) {
-        state.options.onWarning(
-            new YAMLWarning(
-                message,
-                {
-                    column: state.position - state.lineStart,
-                    line: state.line,
-                    position: state.position,
-                },
-                state.input,
-            ),
-        );
-    }
-};
-
-const captureSegment = (state: State, start: number, end: number): void => {
-    if (start < end) {
-        state.result = (state.result as string) + state.input.slice(start, end);
-    }
-};
-
-const writeFoldedLines = (state: State, count: number): void => {
-    if (count === 1) {
-        state.result = `${state.result as string} `;
-    } else if (count > 1) {
-        state.result = (state.result as string) + "\n".repeat(count - 1);
-    }
-};
-
-const readLineBreak = (state: State): void => {
-    const ch = state.input.charCodeAt(state.position);
-
-    if (ch === 0x0a) {
-        state.position++;
-    } else if (ch === 0x0d) {
-        state.position++;
-
-        if (state.input.charCodeAt(state.position) === 0x0a) {
-            state.position++;
-        }
-    } else {
-        throwError(state, "a line break is expected");
-    }
-
-    state.line += 1;
-    state.lineStart = state.position;
-    state.firstTabInLine = -1;
-};
-
-const skipSeparationSpace = (state: State, allowComments: boolean, checkIndent: number): number => {
-    let lineBreaks = 0;
-    let ch = state.input.charCodeAt(state.position);
-
-    while (ch !== 0) {
-        while (isWhiteSpace(ch)) {
-            if (ch === 0x09 && state.firstTabInLine === -1) {
-                state.firstTabInLine = state.position;
-            }
-
-            ch = state.input.charCodeAt(++state.position);
-        }
-
-        if (allowComments && ch === 0x23) {
-            // A `#` only begins a comment when preceded by white space, a line
-            // break, or the start of a line. `]#c`, `"v"#c`, `a,#c` are therefore
-            // malformed rather than commented — this matches the `yaml` reference
-            // (js-yaml is lenient and silently accepts them).
-            const preceding = state.position === state.lineStart ? 0x0a : state.input.charCodeAt(state.position - 1);
-
-            if (!isWsOrEol(preceding)) {
-                throwError(state, "a comment must be separated from other tokens by white space characters");
-            }
-
-            do {
-                ch = state.input.charCodeAt(++state.position);
-            } while (ch !== 0x0a && ch !== 0x0d && ch !== 0);
-        }
-
-        if (isEol(ch)) {
-            readLineBreak(state);
-
-            ch = state.input.charCodeAt(state.position);
-            lineBreaks++;
-            state.lineIndent = 0;
-
-            while (ch === 0x20) {
-                state.lineIndent++;
-                ch = state.input.charCodeAt(++state.position);
-            }
-        } else {
-            break;
-        }
-    }
-
-    if (checkIndent !== -1 && lineBreaks !== 0 && state.lineIndent < checkIndent) {
-        throwError(state, "deficient indentation");
-    }
-
-    return lineBreaks;
-};
-
-const testDocumentSeparator = (state: State): boolean => {
-    let { position } = state;
-    let ch = state.input.charCodeAt(position);
-
-    if ((ch === 0x2d || ch === 0x2e) && ch === state.input.charCodeAt(position + 1) && ch === state.input.charCodeAt(position + 2)) {
-        position += 3;
-        ch = state.input.charCodeAt(position);
-
-        if (ch === 0 || isWsOrEol(ch)) {
-            return true;
-        }
-    }
-
-    return false;
-};
 
 /**
  * Write one mapping key, never letting the document reach the prototype chain.
@@ -488,383 +157,6 @@ const storeMappingPair = (
     keys?.delete(key);
 
     return keys;
-};
-
-const readPlainScalar = (state: State, nodeIndent: number, withinFlowCollection: boolean): boolean => {
-    const previousKind = state.kind;
-    const previousResult = state.result;
-    const { input } = state;
-
-    let ch = input.charCodeAt(state.position);
-
-    if (isWsOrEol(ch) || isFlowIndicator(ch) || isPlainScalarLeadBlocker(ch)) {
-        return false;
-    }
-
-    let following: number;
-
-    if (ch === 0x3f || ch === 0x2d) {
-        following = input.charCodeAt(state.position + 1);
-
-        if (isWsOrEol(following) || (withinFlowCollection && isFlowIndicator(following))) {
-            return false;
-        }
-    }
-
-    state.kind = "scalar";
-    state.result = "";
-
-    let captureStart = state.position;
-    let captureEnd = state.position;
-    let hasPendingContent = false;
-    let line = 0;
-    let lineStart = 0;
-    let lineIndent = 0;
-
-    while (ch !== 0) {
-        if (ch === 0x3a) {
-            following = input.charCodeAt(state.position + 1);
-
-            if (isWsOrEol(following) || (withinFlowCollection && isFlowIndicator(following))) {
-                break;
-            }
-        } else if (ch === 0x23) {
-            const preceding = input.charCodeAt(state.position - 1);
-
-            if (isWsOrEol(preceding)) {
-                break;
-            }
-        } else if ((state.position === state.lineStart && testDocumentSeparator(state)) || (withinFlowCollection && isFlowIndicator(ch))) {
-            break;
-        } else if (isEol(ch)) {
-            line = state.line;
-            lineStart = state.lineStart;
-            lineIndent = state.lineIndent;
-
-            skipSeparationSpace(state, false, -1);
-
-            if (state.lineIndent >= nodeIndent) {
-                hasPendingContent = true;
-                ch = input.charCodeAt(state.position);
-                continue;
-            } else {
-                state.position = captureEnd;
-                state.line = line;
-                state.lineStart = lineStart;
-                state.lineIndent = lineIndent;
-                break;
-            }
-        }
-
-        if (hasPendingContent) {
-            captureSegment(state, captureStart, captureEnd);
-            writeFoldedLines(state, state.line - line);
-            captureStart = state.position;
-            captureEnd = state.position;
-            hasPendingContent = false;
-        }
-
-        if (!isWhiteSpace(ch)) {
-            captureEnd = state.position + 1;
-        }
-
-        ch = input.charCodeAt(++state.position);
-    }
-
-    captureSegment(state, captureStart, captureEnd);
-
-    if (state.result !== "") {
-        return true;
-    }
-
-    state.kind = previousKind;
-    state.result = previousResult;
-
-    return false;
-};
-
-const readSingleQuotedScalar = (state: State, nodeIndent: number): boolean => {
-    let ch = state.input.charCodeAt(state.position);
-
-    if (ch !== 0x27) {
-        return false;
-    }
-
-    state.kind = "scalar";
-    state.result = "";
-    state.position++;
-
-    let captureStart = state.position;
-    let captureEnd = state.position;
-
-    while ((ch = state.input.charCodeAt(state.position)) !== 0) {
-        if (ch === 0x27) {
-            captureSegment(state, captureStart, state.position);
-            ch = state.input.charCodeAt(++state.position);
-
-            if (ch === 0x27) {
-                captureStart = state.position;
-                state.position++;
-                captureEnd = state.position;
-            } else {
-                return true;
-            }
-        } else if (isEol(ch)) {
-            captureSegment(state, captureStart, captureEnd);
-            writeFoldedLines(state, skipSeparationSpace(state, false, nodeIndent));
-            captureStart = state.position;
-            captureEnd = state.position;
-        } else if (state.position === state.lineStart && testDocumentSeparator(state)) {
-            throwError(state, "unexpected end of the document within a single quoted scalar");
-        } else {
-            // Trailing white space before a line break is folded away, so only
-            // advance the captured end past non-space characters (§7.3.1).
-            if (!isWhiteSpace(ch)) {
-                captureEnd = state.position + 1;
-            }
-
-            state.position++;
-        }
-    }
-
-    return throwError(state, "unexpected end of the stream within a single quoted scalar");
-};
-
-const readDoubleQuotedScalar = (state: State, nodeIndent: number): boolean => {
-    let ch = state.input.charCodeAt(state.position);
-
-    if (ch !== 0x22) {
-        return false;
-    }
-
-    state.kind = "scalar";
-    state.result = "";
-    state.position++;
-
-    let captureStart = state.position;
-    let captureEnd = state.position;
-
-    while ((ch = state.input.charCodeAt(state.position)) !== 0) {
-        if (ch === 0x22) {
-            captureSegment(state, captureStart, state.position);
-            state.position++;
-
-            return true;
-        }
-
-        if (ch === 0x5c) {
-            captureSegment(state, captureStart, state.position);
-            ch = state.input.charCodeAt(++state.position);
-
-            if (isEol(ch)) {
-                skipSeparationSpace(state, false, nodeIndent);
-            } else if (ch < 256 && SIMPLE_ESCAPE[ch] !== undefined) {
-                state.result = (state.result as string) + (SIMPLE_ESCAPE[ch] as string);
-                state.position++;
-            } else {
-                const hexLength = escapedHexLength(ch);
-
-                if (hexLength > 0) {
-                    let remaining = hexLength;
-                    let hexResult = 0;
-
-                    for (; remaining > 0; remaining--) {
-                        ch = state.input.charCodeAt(++state.position);
-
-                        const digit = fromHexCode(ch);
-
-                        if (digit >= 0) {
-                            hexResult = (hexResult << 4) + digit;
-                        } else {
-                            throwError(state, "expected hexadecimal character");
-                        }
-                    }
-
-                    state.result = (state.result as string) + String.fromCodePoint(hexResult);
-                    state.position++;
-                } else {
-                    throwError(state, "unknown escape sequence");
-                }
-            }
-
-            captureStart = state.position;
-            captureEnd = state.position;
-        } else if (isEol(ch)) {
-            captureSegment(state, captureStart, captureEnd);
-            writeFoldedLines(state, skipSeparationSpace(state, false, nodeIndent));
-            captureStart = state.position;
-            captureEnd = state.position;
-        } else if (state.position === state.lineStart && testDocumentSeparator(state)) {
-            throwError(state, "unexpected end of the document within a double quoted scalar");
-        } else {
-            // Trailing white space before a line break is folded away, so only
-            // advance the captured end past non-space characters (§7.3.1).
-            if (!isWhiteSpace(ch)) {
-                captureEnd = state.position + 1;
-            }
-
-            state.position++;
-        }
-    }
-
-    return throwError(state, "unexpected end of the stream within a double quoted scalar");
-};
-
-const readBlockScalar = (state: State, nodeIndent: number): boolean => {
-    let ch = state.input.charCodeAt(state.position);
-    let folding: boolean;
-
-    if (ch === 0x7c) {
-        folding = false;
-    } else if (ch === 0x3e) {
-        folding = true;
-    } else {
-        return false;
-    }
-
-    state.kind = "scalar";
-    state.result = "";
-
-    let chomping = CHOMPING_CLIP;
-    let didReadContent = false;
-    let detectedIndent = false;
-    let textIndent = nodeIndent;
-    let emptyLines = 0;
-    let atMoreIndented = false;
-    // Largest indentation seen on a leading (pre-content) empty line — used only
-    // in strict mode to reject a scalar whose leading blank lines are indented
-    // more than its first content line (YAML 1.2 §8.1.1.1).
-    let leadingEmptyIndent = 0;
-
-    while (ch !== 0) {
-        ch = state.input.charCodeAt(++state.position);
-
-        if (ch === 0x2b || ch === 0x2d) {
-            if (chomping === CHOMPING_CLIP) {
-                chomping = ch === 0x2b ? CHOMPING_KEEP : CHOMPING_STRIP;
-            } else {
-                throwError(state, "repeat of a chomping mode identifier");
-            }
-        } else {
-            const decimal = fromDecimalCode(ch);
-
-            if (decimal >= 0) {
-                if (decimal === 0) {
-                    throwError(state, "bad explicit indentation width of a block scalar; it cannot be less than one");
-                } else if (detectedIndent) {
-                    throwError(state, "repeat of an indentation width identifier");
-                } else {
-                    textIndent = nodeIndent + decimal - 1;
-                    detectedIndent = true;
-                }
-            } else {
-                break;
-            }
-        }
-    }
-
-    if (isWhiteSpace(ch)) {
-        do {
-            ch = state.input.charCodeAt(++state.position);
-        } while (isWhiteSpace(ch));
-
-        if (ch === 0x23) {
-            do {
-                ch = state.input.charCodeAt(++state.position);
-            } while (!isEol(ch) && ch !== 0);
-        }
-    }
-
-    while (ch !== 0) {
-        readLineBreak(state);
-        state.lineIndent = 0;
-        ch = state.input.charCodeAt(state.position);
-
-        while ((!detectedIndent || state.lineIndent < textIndent) && ch === 0x20) {
-            state.lineIndent++;
-            ch = state.input.charCodeAt(++state.position);
-        }
-
-        if (!detectedIndent && state.lineIndent > textIndent) {
-            textIndent = state.lineIndent;
-        }
-
-        if (isEol(ch)) {
-            // Remember the widest leading (pre-content) empty line so strict mode
-            // can reject a first content line indented less than them (§8.1.1.1).
-            if (!detectedIndent && state.lineIndent > leadingEmptyIndent) {
-                leadingEmptyIndent = state.lineIndent;
-            }
-
-            emptyLines++;
-            continue;
-        }
-
-        // Strict mode: a tab sitting in the block scalar's indentation (before
-        // the content column is reached with spaces) is illegal — `foo: |\n\t\n`
-        // is malformed, while `foo: |\n \t\n` (tab as content past one space of
-        // indent) stays valid because there `lineIndent === textIndent`.
-        if (state.options.strict && ch === 0x09 && state.lineIndent < textIndent) {
-            throwError(state, "tab characters must not be used in indentation");
-        }
-
-        // End of the scalar: a genuine dedent, EOF, or a `---`/`...` document
-        // marker at column 0. The last case matters only for a zero-indented
-        // block scalar (textIndent 0), where `lineIndent < textIndent` can never
-        // become true; an indented marker keeps `position > lineStart` and stays
-        // scalar content.
-        const atDocumentMarker = state.position === state.lineStart && testDocumentSeparator(state);
-
-        // Strict mode: if the first content line is indented less than a leading
-        // empty line, those empty lines were over-indented (`>` then `  \n   \n # x`)
-        // — a spec error. Only fires for a real content line (not EOF / marker),
-        // so an all-empty keep-chomped scalar (`|+\n   \n`) is unaffected.
-        if (state.options.strict && ch !== 0 && !atDocumentMarker && !detectedIndent && leadingEmptyIndent > state.lineIndent) {
-            throwError(state, "a block scalar's leading empty lines cannot be indented more than its content");
-        }
-
-        if (state.lineIndent < textIndent || ch === 0 || atDocumentMarker) {
-            if (chomping === CHOMPING_KEEP) {
-                state.result = (state.result as string) + "\n".repeat(didReadContent ? 1 + emptyLines : emptyLines);
-            } else if (chomping === CHOMPING_CLIP && didReadContent) {
-                state.result = `${state.result as string}\n`;
-            }
-
-            break;
-        }
-
-        if (folding) {
-            if (isWhiteSpace(ch)) {
-                atMoreIndented = true;
-                state.result = (state.result as string) + "\n".repeat(didReadContent ? 1 + emptyLines : emptyLines);
-            } else if (atMoreIndented) {
-                atMoreIndented = false;
-                state.result = (state.result as string) + "\n".repeat(emptyLines + 1);
-            } else if (emptyLines === 0) {
-                if (didReadContent) {
-                    state.result = `${state.result as string} `;
-                }
-            } else {
-                state.result = (state.result as string) + "\n".repeat(emptyLines);
-            }
-        } else {
-            state.result = (state.result as string) + "\n".repeat(didReadContent ? 1 + emptyLines : emptyLines);
-        }
-
-        didReadContent = true;
-        detectedIndent = true;
-        emptyLines = 0;
-
-        const captureStart = state.position;
-
-        while (!isEol(ch) && ch !== 0) {
-            ch = state.input.charCodeAt(++state.position);
-        }
-
-        captureSegment(state, captureStart, state.position);
-    }
-
-    return true;
 };
 
 const readBlockSequence = (state: State, nodeIndent: number): boolean => {
@@ -1609,11 +901,14 @@ const composeNodeAtDepth = (state: State, parentIndent: number, nodeContext: num
         if (state.kind === "scalar" && typeof state.result === "string") {
             state.result = resolveScalarValue(state.result);
         }
-    } else if (state.tag !== null && state.tag !== "!" // A scalar node, or an explicit tag on empty content (`!!str` with
+    } else if (
+        state.tag !== null
+        && state.tag !== "!" // A scalar node, or an explicit tag on empty content (`!!str` with
         // nothing after it, common in flow: `{ a: !!str, ... }`). In the empty
         // case the node has no kind yet — treat the content as the empty string
         // so `!!str` → "" and `!!null` → null, matching the core schema.
-        && (state.kind === "scalar" || (!hasContent && state.kind === null))) {
+        && (state.kind === "scalar" || (!hasContent && state.kind === null))
+    ) {
         const raw = typeof state.result === "string" ? state.result : "";
         const applied = resolveExplicitTag(state.tag, raw);
